refactor: split manifest section readers
